DeerDetection/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# D:\Skole\VisualStudio_Workspace\DeerDetection\DeerDetection\processThread.py: 14
Pillow == 9.1.0

# D:\Skole\VisualStudio_Workspace\DeerDetection\DeerDetection\startupsetup.py: 1
PyMsgBox == 1.0.9

# D:\Skole\VisualStudio_Workspace\DeerDetection\DeerDetection\processThread.py: 10
geocoder == 1.38.1

# D:\Skole\VisualStudio_Workspace\DeerDetection\DeerDetection\MQTT_subscriberClient\DeerDetection_MQTT_Subscriber.py: 4,7
jsonschema == 4.4.0

# D:\Skole\VisualStudio_Workspace\DeerDetection\DeerDetection\main.py: 2
# D:\Skole\VisualStudio_Workspace\DeerDetection\DeerDetection\processThread.py: 4
numpy == 1.22.3

# D:\Skole\VisualStudio_Workspace\DeerDetection\DeerDetection\input.py: 1
# D:\Skole\VisualStudio_Workspace\DeerDetection\DeerDetection\main.py: 7
# D:\Skole\VisualStudio_Workspace\DeerDetection\DeerDetection\processThread.py: 3
opencv_python == 4.5.5.64

# D:\Skole\VisualStudio_Workspace\DeerDetection\DeerDetection\input.py: 2
pafy == 0.5.5

# D:\Skole\VisualStudio_Workspace\DeerDetection\DeerDetection\MQTT_subscriberClient\DeerDetection_MQTT_Subscriber.py: 1
# D:\Skole\VisualStudio_Workspace\DeerDetection\DeerDetection\processThread.py: 9
paho_mqtt == 1.6.1

# D:\Skole\VisualStudio_Workspace\DeerDetection\DeerDetection\startupsetup.py: 3
requests == 2.27.1

# D:\Skole\VisualStudio_Workspace\DeerDetection\DeerDetection\input.py: 3
# D:\Skole\VisualStudio_Workspace\DeerDetection\DeerDetection\processThread.py: 6
torch == 1.11.0+cu113
